Removed mkview calls because they are breaking navigation commands in rails.vim
The use of mkview seems to break the navigation commands in rails.vim. Most uses of navigation commands creating those errors: "E345: Can't find file “xxx” in path"

A bugreport in the upstream project indicates that the calls to mkview are cause of the errors (https://github.com/tpope/vim-rails/issues/25).
Removing those fixed the problem for me.