Don't use cmaps for one letter command fixes.
As the vimrc is written the cmaps will change any capital W or Q to
lower case. This has the nasty side effect of breaking any functions
with upper case W or Q (e.g., FixWhitespace).

The correct solution is to use commands. The only side effect here is
that all commands must start with an upper case letter.