windows xp installer .cmd
Since there is no mklink in windows XP and installing the tool that
provides this function is only possible on NTFS filesystems I created
this installer for my colleagues at work.

Probably far from a perfect solution but just so people on old XP boxes
can enjoy the awesomness easily.