CtrlP file search command for Git and Hg repositories
Uses the suggested ctrlp_user_command from the [CtrlP website][1] for
searching files in git and Hg repositories that belong to the git or
mercurial repository and ignore files that are ignored by your SCM tool.

Also applies the suggestion from [here][2] to ensure that we include new
untracked files in the search results.

[1]: http://kien.github.com/ctrlp.vim/
[2]: https://github.com/kien/ctrlp.vim/issues/174