Use https links where possible
I've gone through the entire repository, checking all http:// links to
see if the corresponding https urls would also work. This is especially
important when asking the user to run a script they fetched via curl,
because in the case of a mitm they will have no way at all to detect
that something malicious is going on.